Cumulative commit induced by regression tests results.
1. Do learn data writings into service relations after the query end. But
for the full solution of invalidation messages infinity loop problem we need to
save all aqo data in a memory cache. Flushing the data into the disk need to do at
separate transaction.
2. Fix the bug than aqo modifies path_relids and path_clauses node fields during
learning. It was induced segfaults in the case, than the plan was stored into the plancache.
3. Now AQO add state info to EXPLAIN message in verbose mode only.
4. QueryContext was introduced. Now all the query-related aqo data links to the query by
query environment machinery. It is solves a learning problem if subqueries induced by top-level query.
5. Exclude 'SET STORAGE MAIN' definition for some columns of the aqo service relations. It is made
for regression test passing and may be reverted.
6. AQO tests are modified.
7. Suppress AQO in parallel mode. We made it because AQO writes data into service relations
during a parallel stage and break command counter synchronization.